requirements.txt
flask
flask-cors
python-dotenv
nltk
spacy
pandas
langchain
langchain-core
langchain-mistralai
python-dotenv